#49 관리자 컨트롤러, 서비스 분리
