+++ src/main/java/org/example/omnibecard/service/CardAdminService.java
@@ -1,0 +1,10 @@
+package org.example.omnibecard.service;
+
+import org.example.omnibecard.dto.CardResDto;
+import org.springframework.data.domain.Pageable;
+
+public interface CardAdminService {
+
+    CardResDto.GetCardForAdminPage getCardForAdmin(Pageable pageable);
+
+}
