+++ src/main/java/org/example/omnibecard/service/CardService.java
@@ -12,5 +12,5 @@ public interface CardService {
     void verifyCard(Long memberId, String cardPassword);
     CardResDto.GetCard getCard(Long memberId);
     CardResDto.GetMemberId getMemberId(CardReqDto.GetMemberId getMemberIdDto);
-    CardResDto.GetCardForAdminPage getCardForAdmin(Pageable pageable);
+
 }
